requirements.txt
asgiref==3.4.1
bleach==4.1.0
Django~=4.0.6
django-baton==2.2.3
django-js-asset==1.2.2
django-mptt==0.13.4
django-nested-admin==3.4.0
django-stubs==1.9.0
django-stubs-ext==0.3.1
django-summernote==0.8.20.0
mypy==0.910
mypy-extensions==0.4.3
packaging==21.3
psycopg2-binary==2.9.2
pyparsing==3.0.6
python-monkey-business==1.0.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
types-pytz==2021.3.1
types-PyYAML==6.0.1
typing_extensions==4.0.1
webencodings==0.5.1
django-admin-sortable2~=2.1
django-solo~=2.0.0

phonenumbers~=8.12.51
Telethon~=1.24.0